% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hbrplot.R
\name{hbrplot}
\alias{hbrplot}
\title{Produce a glorious HBR style 2x2 plot}
\usage{
hbrplot(items = LETTERS[1:15], quadrants = c("one", "two", "three",
  "four"), x_labels = c(""), y_labels = c(""))
}
\arguments{
\item{items}{A vector of items you want plotted. Typically a string vector.}

\item{quadrants}{A vector of four quandrant names. You can give more than four, but anything after the fourth one is just ignored. Because, seriously.}

\item{x_labels}{A vector of length two which will be coerced into the x axis label.}

\item{y_labels}{A vector of length two which will be coerced into the y axis label.}
}
\value{
A ggplot2 figure that is in the style of a 2x2 plot.
}
\description{
Produce a glorious HBR style 2x2 plot
}
\examples{
hbrplot(items=c('me','you','Dog Named Boo','truck driver'),
        quadrants=c('love','pet','feed','neuter'),
        x_labels=c('fast','slow'),
        y_labels=c('sadly','happily'))
}
